fix: merge packages into manocorp
