--- packages/localtest-docker/localtest_docker.egg-info/requires.txt
@@ -1,1 +1,0 @@
-docker==7.0.0
