--- packages/localtest-docker/localtest_docker.egg-info/top_level.txt
@@ -1,1 +1,0 @@
-localtest_docker
