+++ packages/manocorp/README.md
