+++ packages/manocorp/manocorp/fastapi/routing/__init__.py
@@ -1,0 +1,1 @@
+from .apisecroute import SecFetchJsonRoute
