+++ packages/manocorp/manocorp_utils.egg-info/PKG-INFO
@@ -1,0 +1,9 @@
+Metadata-Version: 2.1
+Name: manocorp-utils
+Version: 0.0.1
+Summary: Manocorp Utils package.
+Author: Manoel Carvalho
+License: MIT
+Requires-Dist: fastapi>=0.110.1
+
+Hello!
