+++ packages/manocorp/manocorp_utils.egg-info/requires.txt
@@ -1,0 +1,1 @@
+fastapi>=0.110.1
